Move generateHashedUsername() and generateHashedEmail() to the new UserHelper class
